Renaming to be wiki, removed manifest.json
