--- public/html/manifest.json
@@ -1,4 +1,0 @@
-[
-  "info.html",
-  "test.html"
-]
