+++ src/app/wiki/page.tsx
@@ -1,12 +1,12 @@
 // app/info/page.tsx
 import { Metadata } from 'next';
-import HtmlDocumentsClient from '@/app/info/htmlDocumentsClient';
+import HtmlDocumentsClient from '@/app/wiki/htmlDocumentsClient';
 
 export const metadata: Metadata = {
   title: 'Info – Civic Dashboard',
 };
 
-export default function Info() {
+export default function Wiki() {
   // Render a client component that fetches from the browser (relative URLs)
   return (
     <div className="min-h-screen">
